Refactor code structure and organization
