+++ src/main/java/com/progress/tracking/wrapper/udemy/pojo/Result.java
@@ -1,4 +1,4 @@
-package com.progress.tracking.response.udemy;
+package com.progress.tracking.wrapper.udemy.pojo;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
